#102: Active server indicator, Connect button, and disabled states
- Mark the currently active server with a faCircleCheck icon and tooltip
- Show a Connect button for non-active servers that closes the modal and
  switches to the selected server
- Disable Edit and Delete for the active server with explanatory tooltips
  on wrapper spans so tooltips work on disabled buttons
- Extract server switch logic from MenuBarCommandHandlerService into
  UiCommandHandlerService as a new UI_SERVER_SWITCH command, making it
  reusable from both the menu and the manage-servers modal
