+++ packages/app/src/app/components/modals/manage-servers/manage-servers.html
@@ -36,33 +36,69 @@
     <ul class="list-group">
       @for (server of filteredServers(); track server.id) {
         <li class="list-group-item d-flex align-items-center justify-content-between">
-          <div class="overflow-hidden me-2">
-            <div class="fw-medium text-truncate">{{ server.name || server.host }}</div>
-            <div class="text-body-secondary small text-truncate">
-              {{ server.protocol }}://{{ server.host }}:{{ server.port }}
+          <div class="overflow-hidden me-2 d-flex align-items-center gap-2">
+            <div class="overflow-hidden">
+              <div class="fw-medium text-truncate">{{ server.name || server.host }}</div>
+              <div class="text-body-secondary small text-truncate">
+                {{ server.protocol }}://{{ server.host }}:{{ server.port }}
+              </div>
             </div>
+            @if (currentServerId() === server.id) {
+              <fa-icon
+                [icon]="icon.faCircleCheck"
+                class="text-success flex-shrink-0"
+                [ngbTooltip]="'components.modals.manage-servers.tooltip.active' | translate"
+                container="body"
+              ></fa-icon>
+            }
           </div>
-          <div class="d-flex gap-1 flex-shrink-0">
-            <button
-              type="button"
-              class="btn btn-link p-1"
-              [ngbTooltip]="'general.button.edit' | translate"
-              [disableTooltip]="editing()"
-              (click)="openEditor(server.id)"
-              [disabled]="editing()"
-            >
-              <fa-icon [icon]="icon.faPencil" />
-            </button>
-            <button
-              type="button"
-              class="btn btn-link text-danger p-1"
-              [ngbTooltip]="'general.button.delete' | translate"
-              [disableTooltip]="editing()"
-              (click)="delete(server)"
-              [disabled]="editing()"
-            >
-              <fa-icon [icon]="icon.faTrashCan" />
-            </button>
+
+          <div class="d-flex gap-1 flex-shrink-0 align-items-center">
+            @if (currentServerId() !== server.id) {
+              <button
+                type="button"
+                class="btn btn-sm btn-outline-primary"
+                (click)="switchTo(server)"
+                [disabled]="editing()"
+              >
+                {{ 'components.modals.manage-servers.button.connect' | translate }}
+              </button>
+              <button
+                type="button"
+                class="btn btn-link p-1"
+                [ngbTooltip]="'general.button.edit' | translate"
+                (click)="openEditor(server.id)"
+                [disabled]="editing()"
+              >
+                <fa-icon [icon]="icon.faPencil" />
+              </button>
+              <button
+                type="button"
+                class="btn btn-link text-danger p-1"
+                [ngbTooltip]="'general.button.delete' | translate"
+                (click)="delete(server)"
+                [disabled]="editing()"
+              >
+                <fa-icon [icon]="icon.faTrashCan" />
+              </button>
+            } @else {
+              <span
+                [ngbTooltip]="'components.modals.manage-servers.tooltip.cannot-edit' | translate"
+                container="body"
+              >
+                <button type="button" class="btn btn-link p-1" disabled>
+                  <fa-icon [icon]="icon.faPencil" />
+                </button>
+              </span>
+              <span
+                [ngbTooltip]="'components.modals.manage-servers.tooltip.cannot-delete' | translate"
+                container="body"
+              >
+                <button type="button" class="btn btn-link text-danger p-1" disabled>
+                  <fa-icon [icon]="icon.faTrashCan" />
+                </button>
+              </span>
+            }
           </div>
         </li>
       }
